books

+++ app/Http/Controllers/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use App\Models\Book;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
@@ -92,7 +93,18 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->books()->count()) {
+            return 'Negalima.';
+        }
         $category->delete();
         return redirect()->route('c_index');
     }
+
+    public function destroyAll(Category $category)
+    {
+        $ids = $category->books()->pluck('id')->all();
+        Book::destroy($ids);
+        return redirect()->route('c_index');
+    }
+
 }
